myproject1/requirements.txt
django
djangorestframework
pyjwt
celery
django-celery-results
django-celery-beat
redis
waitress
djangorestframework-simplejwt